requirements.txt
pillow>=10.0.0
qrcode[pil]>=7.4.2
numpy>=1.24.0